moved the test utils test to controller utisl
